Update client
added the player guessing the word and checking if letters are in the correct place and if they are in the word as a whole, put place holder comments for checking if the letter is in the word but in the wrong spot, as well as a place where it adds the letter to a string that holds letters that are not in the word to guess. Made fixes to opening the file to make sure the integer holding the number of letters wanted in the word to guess is converted to a string before loading which number of letter txt file.