Finished Main Game
Finished Main Game of Server. 
Reverted some changes made in previous commit as they were read incorrectly.
The server and client now completely connect with each other and a full game can be played.
Players can now fully choose the letter amount they would like to play with and the server will choose a word for them.
The words cycle with each game played. If two users connect to the server and choose the same letter length, they will receive a different word from the database.
The word selection is not randomized therefore making it so that the player can not receive the same word twice (even with the very low risk, we can't risk that in a word guessing game).
Maximum of 6 guesses per playthrough has been implemented. May make it so that the server can change that at any time without ever having to update the client. This would allow users to configure their servers in any way they desire and the client (main game) doesn't need to be updated.

TODO:
- Add IP connection to Client so outside clients can join the server.
- Allow player to play multiple times in one sitting rather than reopening the client.
- Score tracking etc.